+++ src/main/java/ee/kool/kool/movies/dto/Film.java
@@ -3,22 +3,22 @@ package ee.kool.kool.movies.dto;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import ee.kool.kool.actors.dto.Actor;
 import lombok.Data;
+import org.springframework.data.annotation.Id;
 
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.ManyToMany;
 import java.util.ArrayList;
 import java.util.List;
 @Data
 public class Film {
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY) private int id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY) private Long id;
     private String name;
     @JsonIgnoreProperties("films")
     @ManyToMany(mappedBy = "films") List<Actor> actors = new ArrayList<>();
 
-    public Film (int id, String name, List<Film> films) {
+    public Film (Long id, String name, List<Film> films) {
         this.id = id;
         this.name = name;
         this.actors = actors;
